commit from local

+++ modules/user/user_api.py
@@ -1,8 +1,8 @@
-from libraries.connect_database import connect_database, User, Product, Cart, Favorite
+from libraries.connect_database import connect_database, User
 from flask_restful import Resource
 from flask import request, jsonify, make_response
-from libraries.libraries import get_user_by_id, get_carts, get_favorites
-from sqlalchemy import exc
+from libraries.libraries import get_user_by_id
+from werkzeug.security import generate_password_hash
 
 session = connect_database()
 
@@ -15,29 +15,23 @@ class SignIn(Resource):
         data = request.get_json()
         user = self.session.query(User).filter_by(username=data['username']).first()
         if user and user.check_password(data['password']):
-            return make_response(
-                jsonify(
-                    {
-                        "message": "done",
-                        "data": {
-                            "id": user.id,
-                            "username": user.username,
-                            "is_admin": user.is_admin,
-                            "carts": get_carts(self.session, Cart, Product, user.id),
-                            "favorites": get_favorites(self.session, Favorite, Product, user.id)
-                        }
-                    }
-                ), 200
-            )
+            user = get_user_by_id(self.session, user.id)
+            message = 'done'
+            data = user
+            code = 200
         else:
-            return make_response(
-                jsonify(
-                    {
-                        "message": "fail",
-                        "data": {}
-                    }
-                ), 401
-            )
+            message = 'fail'
+            data = None
+            code = 401
+        self.session.close()
+        return make_response(
+            jsonify(
+                {
+                    "message": message,
+                    "data": data
+                }
+            ), code
+        )
 
 
 class SignUp(Resource):
@@ -51,25 +45,20 @@ class SignUp(Resource):
             user = User(data['username'], data['password'])
             self.session.add(user)
             self.session.commit()
-            self.session.close()
-            return make_response(
-                jsonify(
-                    {
-                        "message": "done",
-                        "data": {}
-                    }
-                ), 200
-            )
+            message = 'done'
+            code = 200
         else:
-            self.session.close()
-            return make_response(
-                jsonify(
-                    {
-                        "message": "fail",
-                        "data": {}
-                    }
-                ), 401
-            )
+            message = 'fail'
+            code = 401
+        self.session.close()
+        return make_response(
+            jsonify(
+                {
+                    "message": message,
+                    "data": {}
+                }
+            ), code
+        )
 
 
 class UserAPI(Resource):
@@ -77,7 +66,7 @@ class UserAPI(Resource):
         self.session = session()
 
     def get(self, user_id):
-        user = get_user_by_id(self.session, User, user_id)
+        user = get_user_by_id(self.session, user_id)
         self.session.close()
         return make_response(
             jsonify(
@@ -90,27 +79,57 @@ class UserAPI(Resource):
 
     def put(self, user_id):
         data = request.get_json()
-        self.session.query(User).filter_by(id=user_id).update(data)
-        self.session.commit()
-        user = get_user_by_id(self.session, User, user_id)
+        if 'current_password' in data and 'new_password' in data:
+            user = self.session.query(User).filter_by(id=user_id).first()
+            if user.check_password(data['current_password']):
+                self.session.query(User).filter_by(id=user_id).update(
+                    {
+                        'password': generate_password_hash(data['new_password'])
+                    }
+                )
+                self.session.commit()
+                user = get_user_by_id(self.session, user_id)
+                message = 'done'
+                code = 200
+            else:
+                message = 'fail'
+                user = None
+                code = 401
+        else:
+            self.session.query(User).filter_by(id=user_id).update(data)
+            self.session.commit()
+            user = get_user_by_id(self.session, user_id)
+            message = 'done'
+            code = 200
         self.session.close()
         return make_response(
             jsonify(
                 {
-                    "message": "done",
+                    "message": message,
                     "user": user
                 }
-            ), 200
-        )            
+            ), code
+        )
 
     def delete(self, user_id):
-        self.session.query(User).filter_by(id=user_id).delete()
-        self.session.commit()
+        data = request.get_json()
+        user = self.session.query(User).filter_by(id=user_id).first()
+        if user.check_password(data['password']):
+            self.session.query(User).filter_by(id=user_id).delete()
+            self.session.commit()
+            message = 'done'
+            code = 200
+        else:
+            message = 'fail'
+            code = 401
+            self.session.close()
+            
+        self.session.close()
         return make_response(
             jsonify(
                 {
-                    "message": "done",
+                    "message": message,
                     "data": None
                 }
-            ), 200
+            ), code
         )
